refactor(direct-control): extract map read wrappers
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only the map summary, plot snapshot, and map grid read orchestration plus their command builders, embedded plot snapshot source, field normalization, and bounded-grid input validation into packages/civ7-direct-control/src/play/map/reads.ts while keeping public facade exports in index.ts.

Preserves the read-only behavior boundary: map summary still uses executeCiv7Command with the tuner state default, parse labels remain unchanged, map-grid bounds and hiddenInfoPolicy semantics remain local-equivalent, and visibility summary, reveal mutation, GameInfo rows, setup map rows, and player/unit/city summaries remain pending separate owner slices.

Verification:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- map-and-visibility.test.ts
- bun run --cwd packages/civ7-direct-control build
- bun run --cwd packages/cli test -- game.control.test.ts -t "bounded map|GameInfo|AI levers"
- bun run --cwd packages/civ7-direct-control test
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- git diff --check
- bun run check:cli
- bun run test:cli:play
- ownership/no-cycle rg scans for map read ownership, facade type-only imports, and excluded visibility/reveal/GameInfo scope

Local read-only package/source relocation proof only; no runtime or live-game proof claimed.
